Localization fixes (#2)
* Remove unnecessary localize function, use hass.localize() instead.

* Convert scales based in C to F if HA's unit system is set to US customary.

Doing a straight up conversion for now. There could well be reasons for
why it'd be nice to have a separate scale with the stops better aligned,
but it'd increase the maintenance burden for scales a bit. will consider
revisiting this if a use case comes up.

* Remove the default unit definition for `volume` when querying recorder

This was a remnant of aping what the Energy dashboard was doing, can't
think of a good reason to keep it. I also don't think that there'll be
any need for absolute scales featuring volume, so m3 or ft3 shouldn't
really matter to us in the same way that we need to care about C/F.